"VERS" is not fixed-length.
We don't need any definitions for this, but previous airframes failed to
byte-swap the length field for VERS messages, so kludging in a constant
size "fixed" them.

Now the version string has a different length and so "fixing" the length
to the old constant actually desynchronizes the log.

If you want to read data from Launch-11, we'll need a better fix.